Move update DB function to boss_db module.
